requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
emoji==2.8.0
exceptiongroup==1.1.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Levenshtein==0.21.1
magic-filter==1.0.11
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
python-Levenshtein==0.21.1
python-telegram-bot==20.4
pytz==2023.3
rapidfuzz==3.2.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
tomli==2.0.1
types-requests==2.31.0.1
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
